+++ src/MySQL/MySQLFieldSqlBuilder.php
@@ -5,6 +5,7 @@ namespace Wikibase\Database\MySQL;
 use RuntimeException;
 use Wikibase\Database\Escaper;
 use Wikibase\Database\Schema\Definitions\FieldDefinition;
+use Wikibase\Database\Schema\Definitions\TypeDefinition;
 use Wikibase\Database\Schema\FieldSqlBuilder;
 
 /**
@@ -61,29 +62,30 @@ class MySQLFieldSqlBuilder extends FieldSqlBuilder {
 	/**
 	 * Returns the MySQL field type for a given FieldDefinition type constant.
 	 *
-	 * @param string $fieldType
+	 * @param TypeDefinition $fieldType
 	 *
 	 * @return string
 	 * @throws RuntimeException
 	 */
 	protected function getFieldType( $fieldType ) {
-		switch ( $fieldType ) {
+		$fieldTypeName = $fieldType->getName();
+		switch ( $fieldTypeName ) {
 			// No datatype for short strings, i.e. VARCHAR? TEXT or BLOB fields should not be used for that.
-			case FieldDefinition::TYPE_INTEGER:
+			case TypeDefinition::TYPE_INTEGER:
 				return 'INT';
-			case FieldDefinition::TYPE_DECIMAL:
+			case TypeDefinition::TYPE_DECIMAL:
 				return 'DECIMAL';
-			case FieldDefinition::TYPE_BIGINT:
+			case TypeDefinition::TYPE_BIGINT:
 				return 'BIGINT';
-			case FieldDefinition::TYPE_FLOAT:
+			case TypeDefinition::TYPE_FLOAT:
 				return 'FLOAT';
 			//todo define max length of blob fields?
-			case FieldDefinition::TYPE_BLOB:
+			case TypeDefinition::TYPE_BLOB:
 				return 'BLOB'; // This is 64k max.
-			case FieldDefinition::TYPE_TINYINT:
+			case TypeDefinition::TYPE_TINYINT:
 				return 'TINYINT';
 			default:
-				throw new RuntimeException( __CLASS__ . ' does not support db fields of type ' . $fieldType );
+				throw new RuntimeException( __CLASS__ . ' does not support db fields of type ' . $fieldTypeName );
 		}
 	}