+++ tests/phpunit/MySQL/MySQLSchemaSqlBuilderTest.php
@@ -5,6 +5,7 @@ namespace Wikibase\Database\Tests\MySQL;
 use Wikibase\Database\MySQL\MySQLSchemaSqlBuilder;
 use Wikibase\Database\Schema\Definitions\FieldDefinition;
 use Wikibase\Database\Schema\Definitions\IndexDefinition;
+use Wikibase\Database\Schema\Definitions\TypeDefinition;
 use Wikibase\Database\Tests\TestDoubles\Fakes\FakeTableNameFormatter;
 
 /**
@@ -48,7 +49,7 @@ class MySQLSchemaSqlBuilderTest extends \PHPUnit_Framework_TestCase {
 
 	public function testGetAddFieldSql(){
 		$instance = $this->newInstance();
-		$field = new FieldDefinition( 'intField', FieldDefinition::TYPE_INTEGER, FieldDefinition::NOT_NULL, 42 );
+		$field = new FieldDefinition( 'intField', new TypeDefinition( TypeDefinition::TYPE_INTEGER ), FieldDefinition::NOT_NULL, 42 );
 		$sql = $instance->getAddFieldSql( 'tableName', $field );
 		$this->assertEquals( 'ALTER TABLE -prefix_tableName- ADD -intField- INT DEFAULT |42| NOT NULL', $sql );
 	}