+++ tests/phpunit/SQLite/SQLiteFieldSqlBuilderTest.php
@@ -3,6 +3,7 @@
 namespace Wikibase\Database\Tests\SQLite;
 
 use Wikibase\Database\Schema\Definitions\FieldDefinition;
+use Wikibase\Database\Schema\Definitions\TypeDefinition;
 use Wikibase\Database\SQLite\SQLiteFieldSqlBuilder;
 
 /**
@@ -24,7 +25,15 @@ class SQLiteFieldSqlBuilderTest extends \PHPUnit_Framework_TestCase {
 	public function testGetCreateTableSql( FieldDefinition $field, $expectedSQL ) {
 		$mockEscaper = $this->getMock( 'Wikibase\Database\Escaper' );
 
-		$mockEscaper->expects( $this->exactly( $field->getDefault() === null ? 0 : 1 ) )
+		//Dont expect quoted Ints!
+		$fieldTypeName = $field->getType()->getName();
+		if( $fieldTypeName === TypeDefinition::TYPE_INTEGER || $fieldTypeName === TypeDefinition::TYPE_BIGINT || $fieldTypeName === TypeDefinition::TYPE_TINYINT ){
+			$expectedEscapedValue = 0;
+		} else {
+			$expectedEscapedValue = $field->getDefault() === null ? 0 : 1;
+		}
+
+		$mockEscaper->expects( $this->exactly( $expectedEscapedValue ) )
 			->method( 'getEscapedValue' )
 			->will( $this->returnCallback( function( $value ) {
 				return '|' . $value . '|';
@@ -48,7 +57,7 @@ class SQLiteFieldSqlBuilderTest extends \PHPUnit_Framework_TestCase {
 		$argLists[] = array(
 			new FieldDefinition(
 				'fieldName',
-				FieldDefinition::TYPE_TINYINT
+				new TypeDefinition( TypeDefinition::TYPE_TINYINT )
 			),
 			'-fieldName- TINYINT NULL'
 		);
@@ -56,10 +65,9 @@ class SQLiteFieldSqlBuilderTest extends \PHPUnit_Framework_TestCase {
 		$argLists[] = array(
 			new FieldDefinition(
 				'autoInc',
-				FieldDefinition::TYPE_INTEGER,
+				new TypeDefinition( TypeDefinition::TYPE_INTEGER ),
 				FieldDefinition::NOT_NULL,
 				FieldDefinition::NO_DEFAULT,
-				FieldDefinition::NO_ATTRIB,
 				FieldDefinition::AUTOINCREMENT
 
 			),
@@ -69,10 +77,9 @@ class SQLiteFieldSqlBuilderTest extends \PHPUnit_Framework_TestCase {
 		$argLists[] = array(
 			new FieldDefinition(
 				'autoInc',
-				FieldDefinition::TYPE_BIGINT,
+				new TypeDefinition( TypeDefinition::TYPE_BIGINT ),
 				FieldDefinition::NOT_NULL,
 				FieldDefinition::NO_DEFAULT,
-				FieldDefinition::NO_ATTRIB,
 				FieldDefinition::AUTOINCREMENT
 
 			),
@@ -82,17 +89,17 @@ class SQLiteFieldSqlBuilderTest extends \PHPUnit_Framework_TestCase {
 		$argLists[] = array(
 			new FieldDefinition(
 				'fieldName',
-				FieldDefinition::TYPE_TINYINT,
+				new TypeDefinition( TypeDefinition::TYPE_TINYINT ),
 				FieldDefinition::NOT_NULL,
 				'1'
 			),
-			'-fieldName- TINYINT DEFAULT |1| NOT NULL'
+			'-fieldName- TINYINT DEFAULT 1 NOT NULL'
 		);
 
 		$argLists[] = array(
 			new FieldDefinition(
 				'fieldName',
-				FieldDefinition::TYPE_BLOB,
+				new TypeDefinition( TypeDefinition::TYPE_BLOB ),
 				FieldDefinition::NOT_NULL,
 				'foo'
 			),
@@ -105,7 +112,7 @@ class SQLiteFieldSqlBuilderTest extends \PHPUnit_Framework_TestCase {
 	public function testUnsupportedType() {
 		$this->setExpectedException( 'RuntimeException', 'does not support db fields of type' );
 		$sqlBuilder = new SQLiteFieldSqlBuilder( $this->getMock( 'Wikibase\Database\Escaper' ) );
-		$sqlBuilder->getFieldSQL( new FieldDefinition( 'fieldName', 'foobar' ) );
+		$sqlBuilder->getFieldSQL( new FieldDefinition( 'fieldName', new TypeDefinition( 'sdfgsdfgsdfg' ) ) );
 	}
 
 }