+++ tests/phpunit/Schema/Definition/FieldDefinitionTest.php
@@ -4,6 +4,7 @@ namespace Wikibase\Database\Tests\Schema\Definition;
 
 use ReflectionClass;
 use Wikibase\Database\Schema\Definitions\FieldDefinition;
+use Wikibase\Database\Schema\Definitions\TypeDefinition;
 
 /**
  * @covers Wikibase\Database\Schema\Definitions\FieldDefinition
@@ -21,52 +22,28 @@ class FieldDefinitionTest extends \PHPUnit_Framework_TestCase {
 
 		$constructorArgs[] = array(
 			'names',
-			FieldDefinition::TYPE_BLOB
-		);
-
-		$constructorArgs[] = array(
-			'numbers',
-			FieldDefinition::TYPE_FLOAT
+			new TypeDefinition( TypeDefinition::TYPE_BLOB )
 		);
 
 		$constructorArgs[] = array(
 			'stuffs',
-			FieldDefinition::TYPE_INTEGER,
+			new TypeDefinition( TypeDefinition::TYPE_INTEGER ),
 			FieldDefinition::NOT_NULL,
-			42,
-			FieldDefinition::ATTRIB_UNSIGNED
-		);
-
-		$constructorArgs[] = array(
-			'bigint stuffs',
-			FieldDefinition::TYPE_BIGINT,
-			FieldDefinition::NOT_NULL,
-			42,
-			FieldDefinition::ATTRIB_UNSIGNED
-		);
-
-		$constructorArgs[] = array(
-			'decimal stuffs',
-			FieldDefinition::TYPE_DECIMAL,
-			FieldDefinition::NOT_NULL,
-			42,
-			FieldDefinition::ATTRIB_UNSIGNED
+			42
 		);
 
 		$constructorArgs[] = array(
 			'stuffs',
-			FieldDefinition::TYPE_INTEGER,
+			new TypeDefinition( TypeDefinition::TYPE_DECIMAL ),
+			FieldDefinition::NULL,
+			FieldDefinition::NO_DEFAULT
+		);
+
+		$constructorArgs[] = array(
+			'stuffs',
+			new TypeDefinition( TypeDefinition::TYPE_BIGINT ),
 			FieldDefinition::NULL,
 			FieldDefinition::NO_DEFAULT,
-			FieldDefinition::NO_ATTRIB
-		);
-
-		$constructorArgs[] = array(
-			'stuffs',
-			FieldDefinition::TYPE_INTEGER,
-			FieldDefinition::NULL,
-			FieldDefinition::NO_DEFAULT,
-			FieldDefinition::NO_ATTRIB,
 			FieldDefinition::AUTOINCREMENT
 		);
 
@@ -111,13 +88,7 @@ class FieldDefinitionTest extends \PHPUnit_Framework_TestCase {
 		);
 
 		$this->assertEquals(
-			array_key_exists( 4, $constructorArgs ) ? $constructorArgs[4] : FieldDefinition::NO_ATTRIB,
-			$field->getAttributes(),
-			'The FieldDefinition attributes are set and obtained correctly'
-		);
-
-		$this->assertEquals(
-			array_key_exists( 5, $constructorArgs ) ? $constructorArgs[5] : FieldDefinition::NO_AUTOINCREMENT,
+			array_key_exists( 4, $constructorArgs ) ? $constructorArgs[4] : FieldDefinition::NO_AUTOINCREMENT,
 			$field->hasAutoIncrement(),
 			'The FieldDefinition autoIncrement is set and obtained correctly'
 		);
@@ -138,7 +109,7 @@ class FieldDefinitionTest extends \PHPUnit_Framework_TestCase {
 	 */
 	public function testInvalidName( $name ) {
 		$this->setExpectedException( 'InvalidArgumentException' );
-		new FieldDefinition( $name, FieldDefinition::TYPE_INTEGER );
+		new FieldDefinition( $name, new TypeDefinition( TypeDefinition::TYPE_INTEGER ) );
 	}
 
 	public static function invalidTypeProvider(){
@@ -148,6 +119,7 @@ class FieldDefinitionTest extends \PHPUnit_Framework_TestCase {
 			array( null ),
 			array( true ),
 			array( new \Exception() ),
+			array( 'adsfdg' ),
 		);
 	}
 
@@ -173,7 +145,7 @@ class FieldDefinitionTest extends \PHPUnit_Framework_TestCase {
 	 */
 	public function testInvalidNull( $null ) {
 		$this->setExpectedException( 'InvalidArgumentException' );
-		new FieldDefinition( 'name', FieldDefinition::TYPE_INTEGER, $null );
+		new FieldDefinition( 'name', new TypeDefinition( TypeDefinition::TYPE_INTEGER ), $null );
 	}
 
 	/**
@@ -183,10 +155,9 @@ class FieldDefinitionTest extends \PHPUnit_Framework_TestCase {
 		$this->setExpectedException( 'InvalidArgumentException' );
 		new FieldDefinition(
 			'name',
-			FieldDefinition::TYPE_INTEGER,
+			new TypeDefinition( TypeDefinition::TYPE_INTEGER ),
 			FieldDefinition::NULL,
 			FieldDefinition::NO_DEFAULT,
-			FieldDefinition::NO_ATTRIB,
 			$autoinc
 		);
 	}